# docs/swagger.yaml
basePath: /
definitions:
...
  controller.alquilerRequest:
    properties:
      cliente_id:
        type: integer
      deposito_garantia:
        example: 1200.5
        type: number
      fecha_fin:
        type: string
      fecha_inicio:
        type: string
      moneda:
        type: string
      monto_renta:
        example: 1200.5
        type: number
      observaciones:
        type: string
      unidad_id:
        type: integer
      vencimiento_dia_pago:
        type: integer
    type: object
  controller.alquilerResponse:
    properties:
      cliente:
        type: string
      cliente_id:
        type: integer
      estado:
        type: string
      fecha_fin:
        type: string
      fecha_inicio:
        type: string
      id:
        type: integer
      moneda:
        type: string
      monto:
        $ref: '#/definitions/money.Amount'
      unidad:
        type: string
      unidad_id:
        type: integer
    type: object
...
  controller.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  controller.listadoAlquileresResponse:
    properties:
      datos:
        items:
          $ref: '#/definitions/controller.alquilerResponse'
        type: array
      paginacion:
        $ref: '#/definitions/controller.paginadorResponse'
    type: object
...
  controller.paginadorResponse:
    properties:
      pagina:
        type: integer
      pagina_actual:
        type: integer
      paginas:
        type: integer
      por_pagina:
        type: integer
      total:
        type: integer
    type: object
  controller.pagoAlquilerResponse:
    properties:
      alquiler_id:
        type: integer
      fecha_pago:
        type: string
      id:
        type: integer
      mes_correspondiente:
        type: integer
      metodo_pago:
        type: string
      moneda:
        type: string
      monto_pagado:
        $ref: '#/definitions/money.Amount'
      nota:
        type: string
      numero_recibo:
        type: string
    type: object
  controller.pendientePagoResponse:
    properties:
      alquiler_id:
        type: integer
      cliente:
        type: string
      estado:
        type: string
      fecha_vencimiento:
        type: string
      monto:
        $ref: '#/definitions/money.Amount'
      unidad:
        type: string
    type: object
  controller.registrarPagoRequest:
    properties:
      alquiler_id:
        type: integer
      fecha_pago:
        type: string
      mes_correspondiente:
        type: integer
      metodo_pago:
        type: string
      monto_pagado:
        example: 1200.5
        type: number
      nota:
        type: string
    type: object
...
  money.Amount:
    type: object
info:
  contact: {}
  description: API para autenticacion, administracion de empresas y catalogos de soporte
    para el frontend de Rentals Go.
  title: Rentals Go API
  version: "1.0"
paths:
...
  /api/user/alquileres:
    get:
      description: Lista contratos de alquiler con filtros y paginación.
      parameters:
      - description: ID de la empresa
        in: query
        name: empresa_id
        required: true
        type: integer
      - description: Buscar por cliente o código de unidad
        in: query
        name: buscar
        type: string
      - description: Estado del contrato
        in: query
        name: estado
        type: string
      - description: ID de la unidad
        in: query
        name: unidad_id
        type: integer
      - description: Página
        in: query
        name: pag
        type: integer
      - description: Tamaño de página
        in: query
        name: por_pagina
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.listadoAlquileresResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.errorResponse'
      security:
      - BearerAuth: []
      summary: Listar alquileres
      tags:
      - Alquileres
    post:
      consumes:
      - application/json
      description: Crea un nuevo contrato y ocupa la unidad.
      parameters:
      - description: Datos del contrato
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.alquilerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.alquilerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.errorResponse'
      security:
      - BearerAuth: []
      summary: Registrar un alquiler
      tags:
      - Alquileres
  /api/user/alquileres/{id}:
    delete:
      parameters:
      - description: ID del alquiler
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Eliminar contrato de alquiler
      tags:
      - Alquileres
    get:
      description: Devuelve el detalle de un contrato de alquiler.
      parameters:
      - description: ID del alquiler
        in: path
        name: id
        required: true
        type: integer
      - description: ID de la empresa
        in: query
        name: empresa_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.alquilerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.errorResponse'
      security:
      - BearerAuth: []
      summary: Obtener alquiler por ID
      tags:
      - Alquileres
    put:
      parameters:
      - description: ID del alquiler
        in: path
        name: id
        required: true
        type: integer
      - description: Datos a actualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.alquilerRequest'
      responses: {}
      summary: Editar contrato de alquiler
      tags:
      - Alquileres
  /api/user/alquileres/{id}/terminar:
    post:
      parameters:
      - description: ID del alquiler
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Finalizar contrato formalmente
      tags:
      - Alquileres
...
  /api/user/pagos:
    get:
      responses: {}
      summary: Historial de pagos
      tags:
      - Pagos
    post:
      consumes:
      - application/json
      description: Registra un cobro para un alquiler.
      parameters:
      - description: Datos del pago
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.registrarPagoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.pagoAlquilerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.errorResponse'
      security:
      - BearerAuth: []
      summary: Registrar pago de alquiler
      tags:
      - Pagos
  /api/user/pagos/{id}:
    delete:
      responses: {}
      summary: Anular un pago
      tags:
      - Pagos
    get:
      responses: {}
      summary: Detalle de un pago
      tags:
      - Pagos
    put:
      parameters:
      - description: ID del pago
        in: path
        name: id
        required: true
        type: integer
      - description: Notas o metodo_pago
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      responses: {}
      summary: Editar notas o método de pago
      tags:
      - Pagos
  /api/user/pagos/pendientes:
    get:
      description: Retorna contratos sin pago completo registrado para el mes actual.
      parameters:
      - description: ID de la empresa
        in: query
        name: empresa_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.pendientePagoResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.errorResponse'
      security:
      - BearerAuth: []
      summary: Listar pagos pendientes del mes actual
      tags:
      - Pagos
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Token JWT en formato: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"